+++ cargo-e/src/e_manifest.rs
@@ -200,6 +200,7 @@ pub fn get_required_features_from_manifest(
     kind: &TargetKind,
     target_name: &str,
 ) -> Option<String> {
+    trace!("Searching for required features in manifest: {}", manifest_path.display());
     // Read and parse the manifest file.
     let content = fs::read_to_string(manifest_path).ok()?;
     let value: Value = content.parse().ok()?;
@@ -236,7 +237,12 @@ pub fn get_required_features_from_manifest(
         // Convert the manifest_path to a &str.
         if let Some(manifest_str) = manifest_path.to_str() {
             if let Ok(members) = collect_workspace_members(manifest_str) {
-                for (_, member_manifest_path) in members {
+                for (member_name, member_manifest_path) in members {
+                    trace!("Checking workspace {}: {}",member_name, member_manifest_path.display());
+                    if member_name == "." {
+                        trace!("Skipping current workspace member: {}", member_manifest_path.display());
+                        continue; // manifest as the member; recursive loop.
+                    }
                     if let Some(feats) = get_required_features_from_manifest(
                         &member_manifest_path,
                         kind,